move create queue flow to check in button
