+++ packages/app/components/Today/QueueCheckInButton.tsx
@@ -1,49 +1,81 @@
 import { API } from "@koh/api-client";
 import { QueuePartial, Role } from "@koh/common";
-import { message, Modal, Select } from "antd";
+import { Button, message } from "antd";
 import { useRouter } from "next/router";
 import React, { ReactElement, useState } from "react";
 import { useCourse } from "../../hooks/useCourse";
 import { useProfile } from "../../hooks/useProfile";
 import { useRoleInCourse } from "../../hooks/useRoleInCourse";
+import QueueCheckInModal from "./QueueCheckInModal";
+import QueueCreateModal from "./QueueCreateModal";
 import TACheckinButton, { CheckinButton } from "./TACheckinButton";
 
 export default function TodayPageCheckinButton(): ReactElement {
+  // state for check in modal
+  const [checkInModalVisible, setCheckInModalVisible] = useState(false);
+  const [createQueueModalVisible, setCreateQueueModalVisible] = useState(false);
+
   const profile = useProfile();
-  const [modalVisible, setModalVisible] = useState(false);
-  const [queueToCheckInto, setQueueToCheckInto] = useState(-1);
   const router = useRouter();
   const { cid } = router.query;
   const { course, mutateCourse } = useCourse(Number(cid));
   const role = useRoleInCourse(Number(cid));
-  const { Option } = Select;
   const queueCheckedIn = course?.queues.find((queue) =>
     queue.staffList.find((staff) => staff.id === profile?.id)
   );
 
-  const onQueueUpdate = (queueIx: number) => {
-    setQueueToCheckInto(queueIx);
+  const numQueues = course?.queues.length;
+
+  const renderProperModal = () => {
+    numQueues !== 0
+      ? setCheckInModalVisible(true)
+      : setCreateQueueModalVisible(true);
+  };
+
+  const submitMakeQueue = async (submittedForm) => {
+    setCreateQueueModalVisible(false);
+    const queueRequest = await submittedForm.validateFields();
+    try {
+      await API.taStatus.makeQueue(
+        Number(cid),
+        queueRequest.officeHourName,
+        !queueRequest.allowTA,
+        queueRequest.notes
+      );
+      message.success(
+        `Created a new queue ${queueRequest.officeHourName}. Checking you in...`
+      );
+      const redirectID = await API.taStatus.checkIn(
+        Number(cid),
+        queueRequest.officeHourName
+      );
+      mutateCourse();
+      router.push(
+        "/course/[cid]/queue/[qid]",
+        `/course/${Number(cid)}/queue/${redirectID.id}`
+      );
+    } catch (err) {
+      message.error(err.response?.data?.message);
+    }
+    setCreateQueueModalVisible(false);
+  };
+
+  const onCreateQueueButtonClick = () => {
+    setCheckInModalVisible(false);
+    setCreateQueueModalVisible(true);
   };
 
   return (
     <>
-      {modalVisible && (
-        <Modal
-          title="Check into an existing queue"
-          visible={modalVisible}
-          onCancel={() => {
-            setModalVisible(false);
-            setQueueToCheckInto(-1);
-          }}
-          okText="Check In"
-          okButtonProps={{ disabled: queueToCheckInto < 0 }}
-          onOk={async () => {
+      {checkInModalVisible && (
+        <QueueCheckInModal
+          visible={checkInModalVisible}
+          onSubmit={async (queueId: number) => {
             let redirectID: QueuePartial;
             try {
-              // ok only enabled on nonnegative values
               redirectID = await API.taStatus.checkIn(
                 Number(cid),
-                course?.queues[queueToCheckInto].room
+                course?.queues[queueId].room
               );
               mutateCourse();
               router.push(
@@ -54,39 +86,29 @@ export default function TodayPageCheckinButton(): ReactElement {
               message.error(err.response?.data?.message);
             }
           }}
-        >
-          <h3>Begin searching for a room</h3>
-          <Select
-            showSearch
-            style={{ width: 200 }}
-            placeholder="Select a course"
-            optionFilterProp="children"
-            onChange={onQueueUpdate}
-            data-cy="select-existing-queue"
-            filterOption={(input, option) => {
-              return (
-                option.children.toLowerCase().indexOf(input.toLowerCase()) >= 0
-              );
-            }}
-          >
-            {course?.queues
-              .filter((q) => (role === Role.TA ? !q.isProfessorQueue : true))
-              .map((q, i) => (
-                <Option
-                  key={i}
-                  value={i}
-                  data-cy={`select-queue-${q.room}`}
-                >{`${q.room}`}</Option>
-              ))}
-          </Select>
-        </Modal>
+          onCancel={() => setCheckInModalVisible(false)}
+          button={
+            <Button onClick={onCreateQueueButtonClick}>Create Queue</Button>
+          }
+          queues={course.queues}
+          role={role}
+        />
+      )}
+      {createQueueModalVisible && (
+        <QueueCreateModal
+          visible={createQueueModalVisible}
+          onSubmit={submitMakeQueue}
+          onCancel={() => setCreateQueueModalVisible(false)}
+          role={role}
+          lastName={profile?.lastName}
+        />
       )}
       {!queueCheckedIn && role !== Role.STUDENT && (
         <CheckinButton
           type="default"
           size="large"
           data-cy="check-in-modal-button"
-          onClick={() => setModalVisible(true)}
+          onClick={() => renderProperModal()}
         >
           Check In
         </CheckinButton>
